% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{sw_test}
\alias{sw_test}
\title{Perform Shapiro-Wilk Test}
\usage{
sw_test(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
value of W statistic
}
\description{
Perform Shapiro-Wilk Test
}
\examples{
sw_test (rnorm (rnorm (100, 0, 1)))
}
